resolve lifetime issues
